pyproject.toml
[project]
name = "social-media-problemcase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docker>=7.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask[async]>=3.0.3",
    "gunicorn>=23.0.0",
    "logto>=0.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.0",
]
